jobScraper/matureJob/requirements.txt
flask
motor
playwright
pymongo
python-dotenv
aiohttp
